% Generated by roxygen2 (4.0.0): do not edit by hand
\name{aveOFave}
\alias{aveOFave}
\title{\code{aveOFave}}
\usage{
aveOFave(x, n = c(3, 5, 10))
}
\arguments{
\item{x}{Univariate xts series.}

\item{n}{Vector of averaging period to use.}
}
\value{
A univariate xts object representing the average of averages.
}
\description{
Returns a xts object representing the average of multiple averages.
}
\examples{
data(data)
RTL:::aveOFave(x=Cl(CL1),n=c(3,5,10))
}
\author{
Philippe Cote <coteph@mac.com>, Nima Safaian <nima.safaian@gmail.com>
}

