% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond.R
\name{bond}
\alias{bond}
\title{\code{bond}}
\usage{
bond(ytm = 0.05, C = 0.05, T2M = 1, m = 2, output = "price")
}
\arguments{
\item{ytm}{Yield to Maturity}

\item{C}{Coupon rate per annum}

\item{T2M}{Time to maturity in years}

\item{m}{Periods per year for coupon payments e.g semi-annual = 2.}

\item{output}{"price", "df" or "duration"}
}
\value{
Price, cash flows data frame and/or duration
}
\description{
Compute bond price, cash flow table and duration
}
\examples{
bond(ytm = 0.05, C = 0.05,T2M = 1,m = 2,output = "price")
bond(ytm = 0.05,C = 0.05,T2M = 1,m = 2,output = "df")
bond(ytm = 0.05,C = 0.05,T2M = 1,m = 2,output = "duration")
}
\author{
Philippe Cote
}
