% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancrudeassayssum}
\alias{cancrudeassayssum}
\title{Summarized data for Canadian crude assays}
\format{
data frame
}
\source{
\url{https://crudemonitor.ca/}
}
\usage{
cancrudeassayssum
}
\description{
Data set with historical Canadian Crude Assays Statistics.
}
\keyword{datasets}
