% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_eia_steo.R
\name{chart_eia_steo}
\alias{chart_eia_steo}
\title{\code{chart_eia_steo}}
\usage{
chart_eia_steo(
  market = "globalOil",
  key = "your EIA.gov API key",
  fig.title = "EIA STEO Global Liquids SD Balance",
  fig.units = "million barrels per day",
  legend.pos = list(x = 0.4, y = 0.53)
)
}
\arguments{
\item{market}{"globalOil" only currently implemented.}

\item{key}{Your private EIA API token.}

\item{fig.title}{Defaults to "EIA STEO Global Liquids SD Balance".}

\item{fig.units}{Defaults to "million barrels per day"}

\item{legend.pos}{Defaults to list(x = 0.4, y = 0.53)}
}
\value{
A plotly object
}
\description{
Supply Demand Balance from EIA Short Term Energy Outlook.
}
\examples{
\dontrun{
chart_eia_steo(key = EIAkey, market = "globalOil")
 }
}
\author{
Philippe Cote
}
