% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chart_Seasonal}
\alias{chart_Seasonal}
\title{\code{chart_Seasonal}}
\usage{
chart_Seasonal(x = Cl(CL1), start_year = 2011,
  end_year = year(Sys.Date()), output = "chart")
}
\arguments{
\item{x}{A univariate xts object to plot. The colnames() will be included in the chart title.}

\item{start_year}{A numeric YYYY start year from the xts object.}

\item{end_year}{A numeric YYYY end year from the xts object.}

\item{output}{"chart" for trellis object or "data" for data frame of values.}
}
\value{
A trellis object if output="chart" or a dataframe if output="data".
}
\description{
Seasonal plot
}
\examples{
data(data)
x=Cl(CL1) ; colnames(x) <- c("WTI")
RTL:::chart_Seasonal(x=x,start_year=2010,end_year=year(Sys.Date()),output="data")
RTL:::chart_Seasonal(x=x,start_year=2010,end_year=year(Sys.Date()),output="chart")
}
\author{
Gordon Evens <gordon.evens@scotiabank.com>,Philippe Cote <coteph@mac.com,philippe.cote@scotiabank.com>
}

