% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CLc2}
\alias{CLc2}
\title{dataset: second line WTI contract OHLC}
\format{
data frame
}
\source{
{CME}
}
\usage{
CLc2
}
\description{
Historical CL 2nd Contract Close adjusted using ohlc object.
}
\keyword{datasets}
