% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.R
\name{cma}
\alias{cma}
\title{\code{cma}}
\usage{
cma(df = df_fut, contract = "CL", yr = 2019, mo = 1)
}
\arguments{
\item{df}{Wide data frame of date, first contract and second contract}

\item{contract}{Tick.prefix in expiry_table e.g. CL for Brent, LCO for Brent}

\item{yr}{Numeric year}

\item{mo}{Number month}
}
\value{
Wide data frame of date and swap prices.
}
\description{
Computes Calendar Month Averages for WTI and Brent and IV for forward starting swaps. Needs to be enhanced for balmo swaps.
}
\examples{
df <- df_fut \%>\% dplyr::filter(series \%in\% c("CL0219","CL0319"))
df <- df \%>\% tidyr::spread(series,value)
cma(df=df,contract="CL",yr=2019,mo=01)
}
\author{
Philippe Cote
}
