% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRReuro.r
\name{CRReuro}
\alias{CRReuro}
\title{\code{CRReuro}}
\usage{
CRReuro(S, X, sigma, r, T, N, type)
}
\arguments{
\item{S}{Stock price.}

\item{X}{Strike price.}

\item{sigma}{Implied volatility e.g. 0.20}

\item{r}{Risk-free rate.}

\item{T}{Time to maturity in years}

\item{N}{Number of time steps. Internally dt = T/N.}

\item{type}{"call" or "put"}
}
\value{
List of asset price tree, option value tree and option price.
}
\description{
European option binomial model on a stock without dividends.For academic purpose only.
Use fOptions::CRRBinomialTreeOptions for real-life usage.
}
\examples{
CRReuro(S=100,X=100,sigma=0.2,r=0.1,T=1,N=5,type="call")
}
\author{
Philippe Cote
}
