% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crudeassaysBP}
\alias{crudeassaysBP}
\title{crudeassaysBP}
\format{
data frame
}
\source{
\url{https://www.bp.com/en/global/bp-global-energy-trading/features-and-updates/technical-downloads/crudes-assays.html}
}
\usage{
crudeassaysBP
}
\description{
Crude Assays from BP.
}
\keyword{datasets}
