% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crudeassaysXOM}
\alias{crudeassaysXOM}
\title{crudeassaysXOM}
\format{
data frame
}
\source{
\url{https://corporate.exxonmobil.com/Crude-oils/Crude-trading/Crude-oil-blends-by-API-gravity-and-by-sulfur-content#APIgravity}
}
\usage{
crudeassaysXOM
}
\description{
Crude Assays from ExxonMobil.
}
\keyword{datasets}
