% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crudepipelines}
\alias{crudepipelines}
\title{crudepipelines}
\format{
data frame
}
\usage{
crudepipelines
}
\description{
Sample GIS data set for North American crude pipelines.
}
\keyword{datasets}
