% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crudes}
\alias{crudes}
\title{Data for crude assays of 50+ types of crude oil.}
\format{
data frame
}
\source{
Canadian Crude Monitor and BP Crude Assays
}
\usage{
crudes
}
\description{
Crude oil qualities.
}
\keyword{datasets}
