% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cushing}
\alias{cushing}
\title{dataset: WTI Cushing Futures and storage utilization}
\format{
list
}
\source{
{CME and EIA}
}
\usage{
cushing
}
\description{
c1, c2, c1c2 and Cushing storage utilization
}
\keyword{datasets}
