% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CushingStocks}
\alias{CushingStocks}
\title{CushingStocks}
\format{
data frame
}
\source{
\url{https://www.eia.gov/opendata/qb.php?sdid=PET.W_EPC0_SAX_YCUOK_MBBL.W}
}
\usage{
CushingStocks
}
\description{
EIA weekly Cushing crude stocks.
}
\keyword{datasets}
