% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cushingStorage}
\alias{cushingStorage}
\title{dataset: Cushing c1c2 and storage utilization}
\format{
data frame
}
\usage{
cushingStorage
}
\description{
Sample dataset for time spreads vs storage utilization.
}
\keyword{datasets}
