% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data_ret}
\alias{data_ret}
\title{\code{data_ret}}
\usage{
data_ret(x, returntype = c("absolute"))
}
\arguments{
\item{x}{Univariate or multivariate xts object.}

\item{returntype}{Define return type: accepts "absolute" or "relative" as input.}
}
\value{
A univariate or multivariate xts object.
}
\description{
Returns a xts object of absolute or relative returns.
}
\examples{
data(data)
RTL:::data_ret(x=CL1,returntype=c("absolute"))
}
\author{
Philippe Cote <coteph@mac.com,philippe.cote@scotiabank.com>
}

