% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_fut}
\alias{df_fut}
\title{df_fut}
\format{
data frame
#' @source \url{https://www.morningstar.com/products/commodities-and-energy}
}
\usage{
df_fut
}
\description{
Futures settlement data set.
}
\keyword{datasets}
