% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dflong}
\alias{dflong}
\title{dataset: commodity prices in a long dataframe format}
\format{
data frame
}
\usage{
dflong
}
\value{
\code{tibble}
}
\description{
Futures settlement data set.
}
\keyword{datasets}
