% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfwide}
\alias{dfwide}
\title{dataset: commodity prices in a wide dataframe format}
\format{
data frame
}
\usage{
dfwide
}
\value{
\code{tibble}
}
\description{
Futures settlement data set.
}
\keyword{datasets}
