% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distdescplot.R
\name{distdescplot}
\alias{distdescplot}
\title{\code{distdescplot}}
\usage{
distdescplot(x = x)
}
\arguments{
\item{x}{Wide dataframe with date column and single series (univariate).}
}
\value{
Multiple plots describing the distribution.
}
\description{
Provides a summary of returns distribution
}
\examples{
x <- dflong \%>\% dplyr::filter(series=="CL01")
x <- returns(df=x,retType="rel",period.return=1,spread=TRUE)
x <- rolladjust(x=x,commodityname=c("cmewti"),rolltype=c("Last.Trade"))
distdescplot(x=x)
}
\author{
Philippe Cote
}
