% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrierSpreadOption.R
\name{.barrierSpreadGreeks}
\alias{.barrierSpreadGreeks}
\title{Analytically calculated Greeks for barrier spread options}
\usage{
.barrierSpreadGreeks(
  F1,
  F2,
  X,
  B,
  sigma1,
  sigma2,
  rho,
  T2M,
  r,
  type,
  barrier_type,
  monitoring
)
}
\description{
Analytically calculated Greeks for barrier spread options
}
\keyword{internal}
