% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficientFrontier.R
\name{efficientFrontier}
\alias{efficientFrontier}
\title{Markowitz Efficient Frontier}
\usage{
efficientFrontier(
  nsims = 5000,
  x = RTL::fizdiffs \%>\% dplyr::select(date, dplyr::contains("WCS")),
  expectedReturns = NULL
)
}
\arguments{
\item{nsims}{Number of portfolio simulations. Defaults to 5000}

\item{x}{List as provided by output of RTL::simMultivariates().}

\item{expectedReturns}{Defaults to NULL using periodic returns means.}
}
\value{
List of portfolios and chart of efficient frontier
}
\description{
Generates random portfolio weights statistics based on absolute returns. Note returns are neither in percentages nor annualized. This function is used for commodities where returns are dollars per units for real assets e.g. storage tanks, pipelines...
}
\examples{
x =  RTL::fizdiffs \%>\% dplyr::select(date, dplyr::contains("WCS"))
efficientFrontier(nsims = 10, x = x, expectedReturns = NULL)
efficientFrontier(nsims = 10, x = x, expectedReturns = c(1,2,3))
}
\author{
Philippe Cote
}
