% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EIAscrape}
\alias{EIAscrape}
\title{\code{EIAscrape}}
\usage{
EIAscrape(seriesid, api_key = api_key)
}
\arguments{
\item{seriesid}{EIA series name.}

\item{api_key}{EIA API token.}
}
\value{
A list of two objects: name as character and data as a data frame.
}
\description{
Fetches raw data from EIA API website.
}
\examples{
data(data)
RTL:::EIAscrape(seriesid="PET.MCRFPUS1.M",api_key="83307AB94D4FC78A247EF9E5FAE67581")
}
\author{
Philippe Cote <coteph@mac.com>, Nima Safaian <nima.safaian@gmail.com>
}

