% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eiaStocks}
\alias{eiaStocks}
\title{eiaStocks}
\format{
data frame
}
\source{
\url{https://www.eia.gov}
}
\usage{
eiaStocks
}
\description{
EIA weekly crude, NG, ULSD and RBOB stocks.
}
\keyword{datasets}
