% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eiaStorageCap}
\alias{eiaStorageCap}
\title{Data for working storage capacity in the US}
\format{
data frame
}
\usage{
eiaStorageCap
}
\description{
EIA working storage capacity in kbs except NG in bcf.
}
\keyword{datasets}
