% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eurodollar}
\alias{eurodollar}
\title{dataset: Eurodollar futures contracts}
\format{
data frame
}
\usage{
eurodollar
}
\value{
\code{tibble}
}
\description{
ED futures contract for December 2024
}
\keyword{datasets}
