% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expiry_table}
\alias{expiry_table}
\title{dataset: expiry of common commodity futures contract.}
\format{
data frame
}
\usage{
expiry_table
}
\value{
\code{tibble}
}
\description{
This dataframe provides detailed information on major futures contracts specifications
pertaining to last settlement, notices and delivery dates. It also provides tickers in some data service.
}
\keyword{datasets}
