% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fetchdata}
\alias{fetchdata}
\title{\code{fetchdata}}
\usage{
fetchdata(comm, term = NULL, type = c("Cl"))
}
\arguments{
\item{comm}{Prefix of xts objects to form a multivariate xts object from or character vector of xts objects to merge.}

\item{term}{If using comm prefix, a numeric vector of suffixes to combine.}

\item{type}{if xts objects are multivariate OHLCV objects, select which OHLCV column to use using quantmod functions (Cl(),Hi(),Lo(),Op(),Vo())}
}
\value{
A multivariate xts object.
}
\description{
Returns a xts object of absolute or relative returns.
}
\examples{
data(data)
RTL:::fetchdata(comm="CL",term=c(1,2),type=c("Cl"))
RTL:::fetchdata(comm=c("CL1","CL2"),term=NULL,type=c("Cl"))
}
\author{
Philippe Cote <coteph@mac.com,philippe.cote@scotiabank.com>
}

