% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{futuresMonths}
\alias{futuresMonths}
\title{dataset: futures contract codes}
\format{
data frame
}
\usage{
futuresMonths
}
\description{
Exchange-traded contract month codes.
}
\keyword{datasets}
