% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{futuresRef}
\alias{futuresRef}
\title{dataset: futures contracts metadata}
\format{
data frame
}
\usage{
futuresRef
}
\description{
Exchange-traded contract month codes and specifications.
}
\keyword{datasets}
