% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fxfwd}
\alias{fxfwd}
\title{dataset: USDCAD FX forward rates}
\format{
list
}
\source{
{Morningstar and https://ca.investing.com/rates-bonds/forward-rates}
}
\usage{
fxfwd
}
\value{
\code{list}
}
\description{
USDCAD historicals and forward curve
}
\keyword{datasets}
