% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fxfwdPoints}
\alias{fxfwdPoints}
\title{dataset: USDCAD FX forward rates}
\format{
data frame
}
\source{
{https://ca.investing.com/rates-bonds/forward-rates}
}
\usage{
fxfwdPoints
}
\description{
USDCAD forward points
}
\keyword{datasets}
