% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoC.R
\name{getBoC}
\alias{getBoC}
\title{Bank of Canada Valet API}
\usage{
getBoC(series)
}
\arguments{
\item{series}{Array of series name: c("FXCADUSD","BD.CDN.2YR.DQ.YLD"). \code{character}}
}
\value{
A long data frame. \code{tibble}
}
\description{
Extracts series from the Bank of Canada's Valet API.
API documentation at https://www.bankofcanada.ca/valet/docs.
}
\examples{
RTL::getBoC(series = c("FXCADUSD","BD.CDN.2YR.DQ.YLD"))
}
\author{
Philippe Cote
}
