% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFutContinuous.R
\name{getFutContinuous}
\alias{getFutContinuous}
\title{\code{getFutContinuous}}
\usage{
getFutContinuous(roots = c("CL"), contracts = c(1:3), from = Sys.Date() - 365)
}
\arguments{
\item{roots}{Contract root codes on Yahoo Finance eg c("CL"), c("CL","HO")}

\item{contracts}{Contracts out eg. c(1:10)}

\item{from}{Date as character string or Date class}
}
\value{
An OHLC long data frame
}
\description{
Extracts Continuous Futures Contracts from Yahoo Finance.
}
\section{Contract root code on Yahoo Finance examples}{

\itemize{
  \item CL = CME WTI
  \item HO = CME ULSD
  \item RB = CME RBOB
  \item BZ = ICE Brent
  \item ...
}
}

\examples{
getFutContinuous(roots = c("CL"), contracts = c(1:3), from = "2021-01-01")
}
\author{
Mikel Buchinski, Joe Rikk, Usman Farooq, Gabriela Sanchez
}
