% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscape.R
\name{getGenscapePipeOil}
\alias{getGenscapePipeOil}
\title{Genscape API call for oil pipelines}
\usage{
getGenscapePipeOil(
  frequency = "daily",
  regions = "Canada",
  pipelineIDs = c(97),
  revision = "revised",
  limit = 5000,
  offset = 0,
  startDate = "2015-01-01",
  endDate = as.character(Sys.Date()),
  apikey = "yourapikey"
)
}
\arguments{
\item{frequency}{"daily" DEFAULT. \code{character}}

\item{regions}{See API webpage. Multiple values separated by commas e.g. "Canada", "GulfCoast"). \code{character}}

\item{pipelineIDs}{See API webpage. c(98,54...) for specific pipes. \code{numeric}}

\item{revision}{See API webpage. \code{character}}

\item{limit}{See API webpage. Max 5000. \code{numeric}}

\item{offset}{See API webpage. \code{numeric}}

\item{startDate}{"yyyy-mm-dd". \code{character}}

\item{endDate}{"yyyy-mm-dd". \code{character}}

\item{apikey}{Your API key. \code{character}}
}
\value{
wide data frame. \code{tibble}
}
\description{
Returns oil pipeline flows in barrels per day data from Genscape API.You need your own credentials.
Refer to API documentation for argument values. It is assumed if you use this function that
you know the pipelines you need to extract to build supply demand balances.
Use the online API to identify the pipeline IDs.
https://developer.genscape.com/docs/services/oil-transportation/operations/GetPipelineFlowValuesV2/
}
\examples{
\dontrun{
getGenscapePipeOil(
  frequency = "daily", regions = "Canada", pipelineIDs = c(97),
  revision = "revised", limit = 5000, offset = 0,
  startDate = "2015-01-01", endDate = as.character(Sys.Date()),
  apikey = "yourapikey"
)
}

}
\author{
Philippe Cote
}
