% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscape.R
\name{getGenscapeStorageOil}
\alias{getGenscapeStorageOil}
\title{\code{getGenscapeStorageOil}}
\usage{
getGenscapeStorageOil(
  regions = "Canada",
  products = "Crude",
  revision = "revised",
  limit = 5000,
  offset = 0,
  apikey = "yourapikey"
)
}
\arguments{
\item{regions}{See API webpage. Multiple values separated by commas e.g. "Canada, Cushing").}

\item{products}{See API webpage. Multiple values separated by commas e.g. "Crude, JetFuel").}

\item{revision}{See API webpage.}

\item{limit}{See API webpage. Max 5000}

\item{offset}{See API webpage.}

\item{apikey}{Your API key as a character string.}
}
\value{
wide data frame
}
\description{
Returns oil storage data from Genscape API.You need your own credentials.
Refer to API documentation for argument values.
https://developer.genscape.com/docs/services/oil-storage/operations/StorageVolumeByOwnerGet
}
\examples{
\dontrun{
getGenscapeStorageOil <- function(regions = "Canada", products = "Crude",
evision = "revised", limit = 5000, offset = 0,
apikey = "<yourapikey>")
}
}
\author{
Philippe Cote
}
