% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morningstar.R
\name{getIRswapCurve}
\alias{getIRswapCurve}
\title{Morningstar Commodities API single call for IR curves}
\usage{
getIRswapCurve(
  currency = "USD",
  from = "2019-01-01",
  iuser = "x@xyz.com",
  ipassword = "pass"
)
}
\arguments{
\item{currency}{Currently only USD LIBOR implemented.}

\item{from}{From date as character string}

\item{iuser}{Morningstar user name as character - sourced locally in examples.}

\item{ipassword}{Morningstar user password as character - sourced locally in examples.}
}
\value{
wide data frame
}
\description{
Extract historical data for tsQuotes in RQuantlib to bootstrap swap curve
using Morningstar and FRED as data source.
}
\examples{
\dontrun{
getIRswapCurve(currency = "USD", from = "2019-08-26", iuser = username, ipassword = password)
}

}
\author{
Philippe Cote
}
