% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holidaysOil}
\alias{holidaysOil}
\title{dataset: NYMEX and ICE holiday calendars}
\format{
data frame
}
\usage{
holidaysOil
}
\value{
\code{tibble}
}
\description{
Holiday calendars for NYMEX and ICE Brent
}
\keyword{datasets}
