% Generated by roxygen2 (4.0.1): do not edit by hand
\name{installgitpackage}
\alias{installgitpackage}
\title{\code{installgitpackage}}
\usage{
installgitpackage(proxy = TRUE, proxyhttp = NULL, packagename = "RTL",
  gitrepo = "risktoollib")
}
\arguments{
\item{proxy}{Logical value. TRUE is proxy is required.}

\item{proxyhttp}{If proxy=TRUE takes the following "http://..."}

\item{packagename}{Name of package to be loaded from Github}

\item{gitrepo}{Name of Github repository}
}
\value{
Installs Github package
}
\description{
Installs packages from Github
}
\examples{
RTL:::installgitpackage(proxy=FALSE,proxyhttp=NULL,packagename="rCharts",gitrepo="ramnathv")
}
\author{
Nima Safaian <nima.safaian@gmail.com>
}

