% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_df_us.R
\name{ir_df_us}
\alias{ir_df_us}
\title{\code{ir_df_us}}
\usage{
ir_df_us(quandlkey = quandlkey, ir.sens = 0.01)
}
\arguments{
\item{quandlkey}{Your Quandl key "quandlkey"}

\item{ir.sens}{Creates plus and minus IR sensitivity scenarios with specified shock value.}
}
\value{
Data frame of zero rates
}
\description{
Extracts US Tresury Zero Rates
}
\examples{
\dontrun{
us.df <- ir_df_us(quandlkey = quandlkey,ir.sens=0.01)
}
}
\author{
Philippe Cote
}
