% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lngterminals}
\alias{lngterminals}
\title{lngterminals}
\format{
data frame
}
\source{
\url{https://www.eia.gov/maps/layer_info-m.php}
}
\usage{
lngterminals
}
\description{
GIS data set for North American LNG terminals.
}
\keyword{datasets}
