% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadEIA}
\alias{loadEIA}
\title{\code{loadEIA}}
\usage{
loadEIA(tickers = tickers, api_key = "83307AB94D4FC78A247EF9E5FAE67581",
  start_date = "2003-01-01", end_date = format(Sys.Date(), "\%Y-\%m-\%d"))
}
\arguments{
\item{tickers}{Dataframe in the same format as the tickers data object in data(data).}

\item{api_key}{EIA token.}

\item{start_date}{Start date as character yyy-mm-dd.}

\item{end_date}{Start date as character yyy-mm-dd.}
}
\value{
A univariate xts object for each tickers.
}
\description{
Returns a xts object representing the average of multiple averages.
}
\examples{
data(data)
RTL:::loadEIA(tickers=subset(tickers,source \%in\% "eia")[10,],api_key="83307AB94D4FC78A247EF9E5FAE67581",start_date="2003-01-01", end_date=format(Sys.Date(),"\%Y-\%m-\%d"))
}
\author{
Nima Safaian <nima.safaian@gmail.com>
}

