% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadEIAbulk}
\alias{loadEIAbulk}
\title{\code{loadEIAbulk}}
\usage{
loadEIAbulk()
}
\value{
Lots of series...
}
\description{
Loads all EIA NG and PET data as a bulk download.
}
\author{
Philippe Cote <coteph@mac.com>, Nima Safaian <nima.safaian@gmail.com>
}

