% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadQuandl}
\alias{loadQuandl}
\title{\code{loadQuandl}}
\usage{
loadQuandl(tickers = subset(tickers, source \%in\% "quandl"),
  token = "yourtoken", start_date = "2003-01-01",
  end_date = format(Sys.Date(), "\%Y-\%m-\%d"))
}
\arguments{
\item{tickers}{Dataframe in the same format as the tickers data object in data(data).}

\item{token}{Quandl token.}

\item{start_date}{Start date as character yyy-mm-dd.}

\item{end_date}{Start date as character yyy-mm-dd.}
}
\value{
A univariate xts object for each tickers.
}
\description{
Returns a xts object representing the average of multiple averages.
}
\examples{
data(data)
RTL:::loadQuandl(tickers=subset(tickers,source \%in\% "quandl")[10,],token="CQzMpZ5DKPkekYQiXSHM",start_date="2003-01-01", end_date=format(Sys.Date(),"\%Y-\%m-\%d"))
}
\author{
Nima Safain <nima.safaian@gmail.com,nima.safaian@scotiabank.com>, Philippe Cote <coteph@mac.com,philippe.cote@scotiabank.com>
}

