% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nghubs}
\alias{nghubs}
\title{nghubs}
\format{
data frame
}
\source{
\url{https://www.eia.gov/maps/layer_info-m.php}
}
\usage{
nghubs
}
\description{
GIS data set for North American NG Trading Hubs.
}
\keyword{datasets}
