% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngpipelines}
\alias{ngpipelines}
\title{ngpipelines}
\format{
data frame
}
\source{
\url{https://www.eia.gov/maps/layer_info-m.php}
}
\usage{
ngpipelines
}
\description{
GIS data set for North American NG pipelines.
}
\keyword{datasets}
