% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngstorage}
\alias{ngstorage}
\title{ngstorage}
\format{
data frame
}
\source{
\url{https://www.eia.gov/maps/layer_info-m.php}
}
\usage{
ngstorage
}
\description{
GIS data set for North American NG storage.
}
\keyword{datasets}
