% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohlc}
\alias{ohlc}
\title{dataset: randomiser to convert settlement into OHLC}
\format{
data frame
}
\source{
{CME}
}
\usage{
ohlc
}
\description{
OHLC profile using historical CL 1st Contract OHLC
}
\keyword{datasets}
