% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
Internal for use of pipe operator. No return value.
}
\description{
Re-exported from dplyr, which re-exports it from magrittr. See \code{?magrittr::pipe} for details.
}
\keyword{internal}
