% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{productsterminals}
\alias{productsterminals}
\title{productsterminals}
\format{
data frame
}
\source{
\url{https://www.eia.gov/maps/layer_info-m.php}
}
\usage{
productsterminals
}
\description{
GIS data set for North American products pipelines.
}
\keyword{datasets}
