% Generated by roxygen2 (4.0.1): do not edit by hand
\name{promptBeta}
\alias{promptBeta}
\title{\code{promptBeta}}
\usage{
promptBeta(comm, term = NULL, period = "all", cmdty = "cmewti",
  betatype = "all")
}
\arguments{
\item{comm}{Commodity prefix for tickers e.g. "NG", "CL". Refer to fetchdata().}

\item{term}{Suffix of commodity tickers or NULL. Refer to fetchdata().}

\item{period}{"all" or numeric period of time in days.}

\item{cmdty}{commodity name in expiry_table to remove returns on roll dates.}

\item{betatype}{"all" "bull" "bear"}
}
\value{
A list of [1] betas as a dataframe and [2] summary statistics of exponential fit if using term=1:n.
}
\description{
Returns betas of multiple xts prices (by using relative returns).
}
\examples{
data(data)
RTL:::promptBeta(comm="CL",term=3,period="all",cmdty="cmewti",betatype="all")
RTL:::promptBeta(comm=c("CL1","CL2","CL3"),term=NULL,period="all",cmdty="cmewti",betatype="all")
}
\author{
Philippe Cote <coteph@mac.com,philippe.cote@scotiabank.com>, NNima Safain <nima.safaian@gmail.com,nima.safaian@scotiabank.com>
}

