% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refineries}
\alias{refineries}
\title{refineries}
\format{
data frame
}
\source{
\url{https://www.eia.gov/maps/layer_info-m.php}
}
\usage{
refineries
}
\description{
GIS data set for North American refineries.
}
\keyword{datasets}
