% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refineryLPdata}
\alias{refineryLPdata}
\title{dataset: refinery LP model sample inputs and outputs}
\format{
list
}
\usage{
refineryLPdata
}
\description{
Simple refinery  to be used in running LP modeling for education purposes.
}
\keyword{datasets}
