% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rolladjust}
\alias{rolladjust}
\title{\code{rolladjust}}
\usage{
rolladjust(x, datatype = c("returns"), commodityname = c("cmewti"),
  rolltype = c("Last.Trade"), secondcontract = NULL, option = FALSE)
}
\arguments{
\item{x}{An xts object of prices or returns.}

\item{datatype}{Type of xts input object: "prices" or "returns".}

\item{commodityname}{Name of commodity in expiry_table. See example below for values.}

\item{rolltype}{Type of contract roll: "Last.Trade" or "First.Notice".}

\item{secondcontract}{Object name of to adjust roll price with if datatype = "prices".}

\item{option}{TRUE if you also want to remove option roll returns. Default is FALSE.}
}
\value{
Roll-adjusted xts object of returns
}
\description{
Returns a xts price or return object adjusted for contract roll. The methodology used to adjust returns is to remove the daily returns on the day after expiry and for prices to adjust historical rolling front month contracts by the size of the roll at each expiry. This is conducive to quantitative trading strategies as it reflects the PL of a financial trader.
}
\examples{
data(data)
unique(expiry_table$cmdty) # for list of commodity names
# To adjust returns
x <- RTL:::data_ret(x=CL1,returntype=c("absolute"))
RTL:::rolladjust(x=x,datatype=c("returns"),commodityname=c("cmewti"),rolltype=c("Last.Trade"),secondcontract=NULL,option=FALSE)
# To adjust prices
RTL:::rolladjust(x=Cl(CL1),datatype=c("prices"),commodityname=c("cmewti"),rolltype=c("Last.Trade"),secondcontract=Cl(CL2),option=FALSE)
}
\author{
Philippe Cote <coteph@mac.com>, Nima Safain <nima.safaian@gmail.com>
}

