% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RollingRegression}
\alias{RollingRegression}
\title{\code{RollingRegression}}
\usage{
RollingRegression(Ra, Rb, width = 12, Rf = 0, attribute = c("Beta",
  "Alpha", "R-Squared"), main = NULL, na.pad = TRUE, ...)
}
\arguments{
\item{Ra}{A univariate xts object of returns.}

\item{Rb}{A univariate or multivariate xts object of returns.}

\item{width}{Number of periods to compute correlation over.}

\item{Rf}{Risk-free rate. Zero by default.}

\item{attribute}{Any of "Beta", "Alpha" or "R-Squared".}

\item{main}{set the chart title, same as in plot.}

\item{na.pad}{TRUE/FALSE If TRUE it adds any times that would not otherwise have been in the result with a value of NA. If FALSE those times are dropped.}

\item{...}{any other passthru parameters.}
}
\value{
A matrix of results.
}
\description{
Returns an xts object of the rolling regression attribute selected.
}
\examples{
data(data)
Ra<-RTL:::data_ret(x=Cl(CL1),returntype=c("relative"))
Rb<-RTL:::data_ret(x=(CL2),returntype=c("relative"))
RTL:::RollingRegression(Ra=Ra, Rb=Rb, width = 20,attribute = c("Beta", "Alpha", "R-Squared"))
}
\author{
Nima Safain <nima.safaian@gmail.com,nima.safaian@scotiabank.com>
}
\references{
Amended version of PerformanceAnalytics package chart.RollingRegression to return data instead of a chart.
}

