% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTL-package.R
\docType{package}
\name{RTL-package}
\alias{RTL}
\alias{RTL-package}
\title{RTL: Risk Tool Library - Trading, Risk, Analytics for Commodities}
\description{
A toolkit for Commodities 'analytics', risk management and trading professionals. Includes functions for API calls to \url{https://www.zema.global/platforms/zema-marketplace}, \url{https://developer.genscape.com/}, and \url{https://www.bankofcanada.ca/valet/docs}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/risktoollib/RTL}
  \item Report bugs at \url{https://github.com/risktoollib/RTL/issues}
}

}
\author{
\strong{Maintainer}: Philippe Cote \email{pcote@ualberta.ca}

Authors:
\itemize{
  \item Philippe Cote \email{pcote@ualberta.ca}
  \item Nima Safaian \email{nima.safaian@gmail.com}
}

}
\keyword{internal}
