% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ry}
\alias{ry}
\title{Sample RBC stock price data set}
\format{
data frame
}
\source{
{Yahoo Finance}
}
\usage{
ry
}
\description{
Adjusted RBC stock price on NYSE with dividend yield.
}
\keyword{datasets}
