% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{steo}
\alias{steo}
\title{dataset: EIA Short Term Energy Outlook}
\format{
plotly object
}
\source{
{eia}
}
\usage{
steo
}
\value{
\code{htmlwidget}
}
\description{
Short Term Energy Outlook from the EIA.
}
\keyword{datasets}
