% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{dataset: Yahoo Finance data sets}
\format{
list
}
\source{
{Yahoo Finance}
}
\usage{
stocks
}
\value{
\code{list}
}
\description{
Traded equity prices and returns
}
\keyword{datasets}
