% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapFutWeight.R
\name{swapFutWeight}
\alias{swapFutWeight}
\title{\code{swapFutWeight}}
\usage{
swapFutWeight(Month = "2020-09-01", contract = "cmewti", exchange = "nymex")
}
\arguments{
\item{Month}{First calendar day of the month.}

\item{contract}{Contract code in data(expiry_table). sort(unique(expiry_table$cmdty)) for options.}

\item{exchange}{Exchange code in data(holidaysOil). Currently only "nymex" and "ice" supported.}
}
\value{
Percentage applied to first contract. To compute swap 1 - value = % applied to 2nd line contract.
}
\description{
Returns the percentage weight of the future in Calendar Month Average swaps
}
\examples{
swapFutWeight(Month = "2020-09-01",contract = "cmewti",exchange = "nymex")
}
\author{
Philippe Cote
}
