% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tickers_eia}
\alias{tickers_eia}
\title{tickers_eia}
\format{data frame}
\source{
\url{https://www.eia.gov/}
}
\usage{
tickers_eia
}
\description{
Supports automated upload of EIA data through its API by categories.
Data frame organized by Supply Demand caterogies and products.
}
\keyword{datasets}
