% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tradeCycle}
\alias{tradeCycle}
\title{dataset: Canadian and US physical crude trading calendars}
\format{
data frame
}
\usage{
tradeCycle
}
\description{
Crude Trading Trade Cycles
}
\keyword{datasets}
