% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tradeprocess}
\alias{tradeprocess}
\title{tradeprocess}
\format{
data frame
}
\usage{
tradeprocess
}
\description{
Data set for explaning the various ways to monetize a market view.
}
\keyword{datasets}
