% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeStats.R
\name{tradeStats}
\alias{tradeStats}
\title{Risk-reward statistics for quant trading}
\usage{
tradeStats(x, Rf = 0)
}
\arguments{
\item{x}{Univariate xts object of returns OR dataframe with date and return variables.}

\item{Rf}{Risk-free rate}
}
\value{
List of risk/reward metrics.
}
\description{
Compute list of risk reward metrics
}
\examples{
library(PerformanceAnalytics)
tradeStats(x = stocks$spy, Rf = 0)
}
\author{
Philippe Cote
}
