% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeStrategyDY.R
\name{tradeStrategyDY}
\alias{tradeStrategyDY}
\title{Sample quantitative trading strategy}
\usage{
tradeStrategyDY(data, par1value = 50, par2value = 200)
}
\arguments{
\item{data}{Dataframe of OHLC data e.g. RTL::uso. \code{tibble}}

\item{par1value}{Value of first parameter e.g. short MA. \code{numeric}}

\item{par2value}{Value of second parameter e.g. long MA. \code{numeric}}
}
\value{
Dataframe with indicators, signals, trades and profit and loss. \code{tibble}
}
\description{
Based on dividend yield.
}
\examples{
tradeStrategyDY(data = RTL::stocks$ry, par1value = 50, par2value = 200)
}
\author{
Philippe Cote
}
