% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeStrategySMA.R
\name{tradeStrategySMA}
\alias{tradeStrategySMA}
\title{Sample quantitative trading strategy}
\usage{
tradeStrategySMA(data = RTL::stocks$uso, par1value = 50, par2value = 200)
}
\arguments{
\item{data}{Dataframe of OHLC data e.g. RTL::uso. \code{tibble}}

\item{par1value}{Value of first parameter e.g. short MA. \code{numeric}}

\item{par2value}{Value of second parameter e.g. long MA. \code{numeric}}
}
\value{
Dataframe with indicators, signals, trades and profit and loss. \code{tibble}
}
\description{
Moving average crossover strategy
}
\examples{
tradeStrategySMA(data = RTL::stocks$uso, par1value = 50, par2value = 200)
}
\author{
Philippe Cote
}
