% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsQuotes}
\alias{tsQuotes}
\title{Interest Rate Curve Data for RQuantlib .}
\format{
data frame
}
\usage{
tsQuotes
}
\description{
USD IR curve input for RQuantlib::DiscountCurve
}
\keyword{datasets}
