% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usSwapCurves}
\alias{usSwapCurves}
\title{dataset: US bootstrapped interest rate curve.}
\format{
List
}
\usage{
usSwapCurves
}
\value{
\code{list}
}
\description{
USD IR Discount, Forward and Zero curves from RQuantlib::DiscountCurve
}
\keyword{datasets}
