% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usSwapCurvesPar}
\alias{usSwapCurvesPar}
\title{usSwapCurvesPar}
\format{data frame}
\usage{
usSwapCurvesPar
}
\description{
USD IR Discount, Forward and Zero curves from RQuantlib::DiscountCurve - Parallel toy data set
}
\keyword{datasets}
