% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usSwapIR}
\alias{usSwapIR}
\title{usSwapIR}
\format{
data frame
#' @source Morningstar and FRED
}
\usage{
usSwapIR
}
\description{
USD Interest Rate Swap Curve for RQuantlib bootstrapping. See usSwapIRdef for sources and tickers.
}
\keyword{datasets}
