% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usSwapIRdef}
\alias{usSwapIRdef}
\title{usSwapIRdef}
\format{data frame
#' @source Morningstar and FRED}
\usage{
usSwapIRdef
}
\description{
USD Interest Rate Swap Curve definitions with sources and tickers
}
\keyword{datasets}
