% Generated by roxygen2 (4.0.1): do not edit by hand
\name{varshockedvalues}
\alias{varshockedvalues}
\title{\code{varshockedvalues}}
\usage{
varshockedvalues(prices, relshocks)
}
\arguments{
\item{prices}{A xts object of prices for the underlying risk factor.}

\item{relshocks}{A xts object of relative returns.}
}
\value{
A xts object of B/A VaR shocks rebased on last price data.
}
\description{
Returns B/A VaR shocks rebased on current prices.
}
\examples{
data(data)
RTL:::varshockedvalues(prices=Cl(CL1),relshocks=RTL:::data_ret(x=Cl(CL1),returntype="relative"))
}
\author{
Philippe Cote <coteph@mac.com,philippe.cote@scotiabank.com>
}

