% Generated by roxygen2 (4.0.0): do not edit by hand
\name{aveOFave}
\alias{aveOFave}
\title{\code{aveOFave}}
\usage{
aveOFave(x, n = c(3, 5, 10))
}
\arguments{
\item{x}{xts}

\item{n}{vector}
}
\value{
a xts object
}
\description{
returns the average of a series of averages specified in vector n.
}

