% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancrudeassays}
\alias{cancrudeassays}
\title{Data for Canadian crude assays reported by Crude Monitor}
\format{
data frame
}
\source{
\url{https://beta.crudemonitor.ca/}
}
\usage{
cancrudeassays
}
\description{
Data set with historical Canadian Crude Assays.
}
\keyword{datasets}
