% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancrudeprices}
\alias{cancrudeprices}
\title{cancrudeprices}
\format{data frame}
\usage{
cancrudeprices
}
\description{
Randomized dataset of Canadian Crude monthly prices versus WTi Calendar Month Average.
}
\keyword{datasets}
