% Generated by roxygen2 (4.0.0): do not edit by hand
\name{chart.RollingCorr}
\alias{chart.RollingCorr}
\title{\code{chart.RollingCorr}}
\usage{
chart.RollingCorr(Ra, Rb, width = 12, xaxis = TRUE, legend.loc = NULL,
  colorset = (1:12), ylimmin = -1, ylimmax = 1, ..., fill = NA)
}
\arguments{
\item{Ra}{A univariate xts object of returns.}

\item{Rb}{A univariate or multivariate xts object of returns.}

\item{width}{Number of periods to compute correlation over.}
}
\value{
A univariate xts object representing the average of averages.
}
\description{
Charts Rolling Correlation
}
\author{
Philippe Cote <coteph@mac.com>, Nima Safain <nima.safaian@gmail.com>
}

