% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_zscore.R
\name{chart_zscore}
\alias{chart_zscore}
\title{\code{chart_zscore}}
\usage{
chart_zscore(
  df = df,
  title = "NG Storage Z Score",
  per = "weekly",
  output = "zscore"
)
}
\arguments{
\item{df}{Long data frame with columns series, date and value}

\item{title}{Your chart title}

\item{per}{Frequency of seasonality "weekly" (DEFAULT) or "monthly"}

\item{output}{"zscore" or "seasonal" chart.}
}
\value{
Time series of STL decomposition residuals Z-Scores, or
standard seasonal chart with feast package.
}
\description{
Provides a summary of returns distribution
}
\examples{
chart_zscore(df = ng_storage, title = "NG Storage Z Score", per = "weekly", output = "zscore")
chart_zscore(df = ng_storage, title = "NG Storage Z Score", per = "weekly", output = "seasonal")
chart_zscore(df = ng_storage, title = "NG Storage Z Score", per = "montlhy", output = "zscore")
}
\author{
Philippe Cote
}
