% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_zscore.R
\name{chart_zscore}
\alias{chart_zscore}
\title{Z-Score applied to seasonal data divergence}
\usage{
chart_zscore(
  df = df,
  title = "NG Storage Z Score",
  per = "yearweek",
  output = "zscore",
  chart = "seasons"
)
}
\arguments{
\item{df}{Long data frame with columns series, date and value. \code{tibble}}

\item{title}{Default is a blank space returning the unique value in df$series. \code{character}}

\item{per}{Frequency of seasonality "yearweek" (DEFAULT). "yearmonth", "yearquarter" \code{character}}

\item{output}{"stl" for STL decomposition chart,
"stats" for STL fitted statistics.
"res" for STL fitted data.
"zscore" for residuals Z-score,
"seasonal" for standard seasonal chart.}

\item{chart}{"seasons" for feasts::gg_season() (DEFAULT)
"series" for feasts::gg_subseries()}
}
\value{
Time series of STL decomposition residuals Z-Scores, or
standard seasonal chart with feast package.
}
\description{
Supports analytics and display of seasonal data. Z-Score is
computed on residuals conditional on their seasonal period.
Beware that most seasonal charts in industry e.g. (NG Storage)
is not de-trended so results once you apply an STL decomposition
will vary from the unajusted seasonal plot.
}
\examples{
\dontrun{
df <- eiaStocks \%>\% dplyr::filter(series == "NGLower48")
title <- "NGLower48"
chart_zscore(df = df, title = " ", per = "yearweek", output = "stl", chart = "seasons")
chart_zscore(df = df, title = " ", per = "yearweek", output = "stats", chart = "seasons")
chart_zscore(df = df, title = " ", per = "yearweek", output = "res", chart = "seasons")
chart_zscore(df = df, title = " ", per = "yearweek", output = "zscore", chart = "seasons")
chart_zscore(df = df, title = " ", per = "yearweek", output = "seasonal", chart = "seasons")
}
}
\author{
Philippe Cote
}
