% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cma}
\alias{cma}
\title{metadata for WTI CMA}
\format{
data frame
}
\source{
{cme}
}
\usage{
cma
}
\value{
\code{tibble}
}
\description{
CME WTI Calendar Month Average swap information
}
\keyword{datasets}
