% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dflong}
\alias{dflong}
\title{dflong}
\format{
data frame
#' @source \url{https://www.morningstar.com/products/commodities-and-energy}
}
\usage{
dflong
}
\description{
Futures settlement data set.
}
\keyword{datasets}
