% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distdescplot.R
\name{distdescplot}
\alias{distdescplot}
\title{Summary of distribution properties of a timeseries}
\usage{
distdescplot(x = x)
}
\arguments{
\item{x}{Wide dataframe with date column and single series (univariate).}
}
\value{
Multiple plots describing the distribution.
}
\description{
Provides a summary of returns distribution
}
\examples{
x <- dplyr::tibble(
  date = seq.Date(Sys.Date() - 1000, Sys.Date(), 1),
  CL01 = c(rnorm(501, 0, 0.02), rnorm(500, 0, 0.01))
)
distdescplot(x = x)
}
\author{
Philippe Cote
}
