% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eiaStorageCap}
\alias{eiaStorageCap}
\title{dataset: EIA working storage capacity}
\format{
data frame
}
\usage{
eiaStorageCap
}
\value{
\code{tibble}
}
\description{
EIA working storage capacity in kbs except NG in bcf.
}
\keyword{datasets}
