% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eiaStorageCap}
\alias{eiaStorageCap}
\title{eiaStorageCap}
\format{
data frame
}
\source{
\url{https://www.eia.gov/petroleum/storagecapacity/}
}
\usage{
eiaStorageCap
}
\description{
EIA crude storage capacity in thousand bbls.
}
\keyword{datasets}
