% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOU.R
\name{fitOU}
\alias{fitOU}
\title{Fits a Ornstein–Uhlenbeck process to a dataset}
\usage{
fitOU(spread)
}
\arguments{
\item{spread}{Spread time series.}
}
\value{
List of alpha, mu and sigma estimates
}
\description{
Parameter estimation for Ornstein–Uhlenbeck process
}
\examples{
spread <- simOU(mu = 5, theta = .5, sigma = 0.2, T = 5, dt = 1 / 250)
fitOU(spread)
}
\author{
Philippe Cote
}
