% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fizdiffs}
\alias{fizdiffs}
\title{fizdiffs}
\format{
data frame
}
\usage{
fizdiffs
}
\description{
Randomized data set for education purpose of selected physical crude differentials to WTI.
}
\keyword{datasets}
