% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fxfwd}
\alias{fxfwd}
\title{Data for USDCAD FX forward rates}
\format{
data frame
}
\source{
{Morningstar}
}
\usage{
fxfwd
}
\description{
USDCAD 1-year and 5-year forward points
}
\keyword{datasets}
