% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurve.R
\name{getCurve}
\alias{getCurve}
\title{\code{getCurve}}
\usage{
getCurve(
  feed = "CME_NymexFutures_EOD",
  contract = "CL",
  from = "2019-08-27",
  iuser = "x@xyz.com",
  ipassword = "pass"
)
}
\arguments{
\item{feed}{Morningstar Feed Table}

\item{contract}{Tick.prefix in expiry_table e.g. CL for Brent, LCO for Brent}

\item{from}{Curve date as character string}

\item{iuser}{Morningstar user name}

\item{ipassword}{Morningstar user password}
}
\value{
wide data
}
\description{
Computes Calendar Month Averages for WTI and Brent and IV for forward starting swaps. Needs to be enhanced for balmo swaps.
}
\examples{
source("~/keys.R")
getCurve(feed="CME_NymexFutures_EOD",contract="CL",
from="2019-08-27",iuser = mstar[[1]], ipassword = mstar[[2]])
}
\author{
Philippe Cote
}
