% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morningstar.R
\name{getCurve}
\alias{getCurve}
\title{\code{getCurve}}
\usage{
getCurve(
  feed = "CME_NymexFuturesIntraday_EOD",
  contract = "CL",
  from = "2020-07-13",
  fields = c("Open, High, Low, Close"),
  iuser = "x@xyz.com",
  ipassword = "pass"
)
}
\arguments{
\item{feed}{Morningstar Feed Table.}

\item{contract}{Morningstar contract root.}

\item{from}{From date as character string.}

\item{fields}{Defaults to c("Open, High, Low, Close").}

\item{iuser}{Morningstar user name as character - sourced locally in examples.}

\item{ipassword}{Morningstar user password as character - sourced locally in examples.}
}
\value{
wide data frame
}
\description{
Returns forward curves from Morningstar API. See below for current feeds supported.
You need your own credentials with Morningstar.
}
\section{Current Feeds Supported}{

\itemize{
  \item CME_NymexFuturesIntraday_EOD
  \item ICE_EuroFutures and ICE_EuroFutures_continuous
}
}

\examples{
\dontrun{
getCurve <- function(feed="CME_NymexFuturesIntraday_EOD",contract="CL",
from="2020-07-13",fields = c("Open, High, Low, Close"),
iuser = "x@xyz.com", ipassword = "pass")
}
}
\author{
Philippe Cote
}
