% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrices.R
\name{getPrices}
\alias{getPrices}
\title{\code{getPrices}}
\usage{
getPrices(
  feed = "CME_NymexFutures_EOD",
  contracts = c("CL9Z", "CL0F", "CL0M"),
  from = "2019-01-01",
  iuser = "x@xyz.com",
  ipassword = "pass"
)
}
\arguments{
\item{feed}{Morningstar Feed Table}

\item{contracts}{Symbols vector}

\item{from}{From date as character string}

\item{iuser}{Morningstar user name}

\item{ipassword}{Morningstar user password}
}
\value{
wide data frame
}
\description{
Returns data from Morningstat API
}
\examples{
source("~/keys.R")
getPrices(feed="CME_NymexFutures_EOD",contracts=c("CL9Z","CL0F","CL0M"),
from="2019-08-26",iuser = mstar[[1]], ipassword = mstar[[2]])
}
\author{
Philippe Cote
}
