% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morningstar.R
\name{getPrices}
\alias{getPrices}
\title{\code{getPrices}}
\usage{
getPrices(
  feed = "CME_NymexFutures_EOD",
  contracts = c("CL9Z", "CL0F", "CL0M"),
  from = "2019-01-01",
  iuser = "x@xyz.com",
  ipassword = "pass"
)
}
\arguments{
\item{feed}{Morningstar Feed Table}

\item{contracts}{Symbols vector}

\item{from}{From date as character string}

\item{iuser}{Morningstar user name as character - sourced locally in examples.}

\item{ipassword}{Morningstar user password as character - sourced locally in examples.}
}
\value{
wide data frame
}
\description{
Multiple Morningstar API calls using getPrice functions.
Refer to `getPrices()` for list of currently supported data feeds.
}
\examples{
\dontrun{
getPrices(feed="CME_NymexFutures_EOD",contracts=c("CL9Z","CL0F","CL0M"),
from="2019-08-26",iuser = username, ipassword = password)
}
}
\author{
Philippe Cote
}
