% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
Internal for use of pipe operator. No return value.
}
\description{
See \verb{magrittr::[\\\%>\\\%][magrittr::pipe]} for details.
}
\keyword{internal}
