% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planets}
\alias{planets}
\title{planets}
\format{
data frame
}
\source{
\url{https://nssdc.gsfc.nasa.gov/planetary/factsheet/index.html}
}
\usage{
planets
}
\description{
Planet metrics from NASA
}
\keyword{datasets}
