% Generated by roxygen2 (4.0.0): do not edit by hand
\name{promptBeta}
\alias{promptBeta}
\title{\code{promptBeta}}
\usage{
promptBeta(comm, term = NULL, period = "all", cmdty = "cmewti",
  betatype = "all")
}
\arguments{
\item{comm}{Commodity prefix for tickers e.g. "NG", "CL". Refer to fetchdata().}

\item{term}{Suffix of commodity tickers or NULL. Refer to fetchdata().}

\item{last}{"all" or numeric period of time in days.}

\item{betatype}{"all" "bull" "bear"}

\item{end_date}{Start date as character yyy-mm-dd.}
}
\value{
A list of [1] betas as a dataframe and [2] summary statistics.
}
\description{
Returns a xts object representing the average of multiple averages.
}
\examples{
data(data)
RTL:::promptBeta(comm="CL",term=3,period="all",cmdty="cmewti",betatype="all")
RTL:::promptBeta(comm=c("CL1","CL2","CL3"),term=NULL,period="all",cmdty="cmewti",betatype="all")
}
\author{
Philippe Cote <coteph@mac.com>, Nima Safaian <nima.safaian@gmail.com>
}

