% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref.opt.inputs}
\alias{ref.opt.inputs}
\title{ref.opt.inputs}
\format{
data frame
}
\usage{
ref.opt.inputs
}
\description{
Simple refinery input to be used in running LP modeling for education purposes.
}
\keyword{datasets}
