% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref.opt.outputs}
\alias{ref.opt.outputs}
\title{Metadata for teaching refinery optimization using a LP model - OUTPUTS}
\format{
data frame
}
\usage{
ref.opt.outputs
}
\description{
Simple refinery outputs and constraints to be used in running LP modeling for education purposes.
}
\keyword{datasets}
