% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineryLP.R
\name{refineryLP}
\alias{refineryLP}
\title{\code{refineryLP}}
\usage{
refineryLP(crudes = ref.opt.inputs, products = ref.opt.outputs)
}
\arguments{
\item{crudes}{Data frame of crude inputs}

\item{products}{Data frame of product outputs and max outputs.}
}
\value{
Optimal crude slate and profits
}
\description{
Plain vanilla refinery optmization LP model.
}
\examples{
refineryLP(crudes = ref.opt.inputs, products = ref.opt.outputs)
}
\author{
Philippe Cote
}
