% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineryLP.R
\name{refineryLP}
\alias{refineryLP}
\title{LP model for refinery optimization}
\usage{
refineryLP(
  crudes = RTL::refineryLPdata$inputs,
  products = RTL::refineryLPdata$outputs
)
}
\arguments{
\item{crudes}{Data frame of crude inputs}

\item{products}{Data frame of product outputs and max outputs.}
}
\value{
Optimal crude slate and profits
}
\description{
Plain vanilla refinery optimization LP model.
}
\examples{
refineryLP(crudes = RTL::refineryLPdata$inputs, products = RTL::refineryLPdata$outputs)
}
\author{
Philippe Cote
}
