% Generated by roxygen2 (4.0.0): do not edit by hand
\name{riskprofile}
\alias{riskprofile}
\title{\code{riskprofile}}
\usage{
riskprofile(x)
}
\arguments{
\item{x}{An xts object of returns.}
}
\value{
A matrix of results
}
\description{
Computes empirical VaR, sVaR and Stress Worst Cases from a returns xts object.
}
\examples{
data(data)
RTL:::riskprofile(x=RTL:::data_ret(x=Cl(CL1),returntype="relative"))
RTL:::riskprofile(x=RTL:::data_ret(x=Cl(CL1),returntype="absolute"))
}
\author{
Philippe Cote <coteph@mac.com>, Nima Safain <nima.safaian@gmail.com>
}

