% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RollingRegression}
\alias{RollingRegression}
\title{\code{RollingRegression}}
\usage{
RollingRegression(Ra, Rb, width = 12, Rf = 0, attribute = c("Beta",
  "Alpha", "R-Squared"), main = NULL, na.pad = TRUE, ...)
}
\arguments{
\item{Ra}{A univariate xts object of returns.}

\item{Rb}{A univariate or multivariate xts object of returns.}

\item{width}{Number of periods to compute correlation over.}

\item{Rf}{Risk-free rate. Zero by default.}

\item{attribute}{Any of "Beta", "Alpha" or "R-Squared".}

\item{...}{any other passthru parameters.}
}
\value{
A matrix of results.
}
\description{
chart.RollingCorrelation from PerformanceAnalytics using Spearman method and customized ylim
}
\examples{
data(data)
Ra<-RTL:::data_ret(x=Cl(CL1),returntype=c("relative"))
Rb<-RTL:::data_ret(x=(CL2),returntype=c("relative"))
RTL:::RollingRegression(Ra=Ra, Rb=Rb, width = 20,attribute = c("Beta", "Alpha", "R-Squared"))
}
\author{
Philippe Cote <coteph@mac.com>, Nima Safain <nima.safaian@gmail.com>
}

