% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMultivariates.R
\name{simMultivariates}
\alias{simMultivariates}
\title{Multivariate Normal from historical dataset}
\usage{
simMultivariates(nsims = 10, x, s0 = NULL)
}
\arguments{
\item{nsims}{Number of simulations. Defaults to 10}

\item{x}{Wide data frame of prices with date as first column.}

\item{s0}{Vector of starting value for each variables. Defaults to NULL with zero.}
}
\value{
List of means, sds, covariance matrix, correlation matrix and simulated values
}
\description{
Generates multivariate random epsilons using absolute returns.
}
\examples{
simMultivariates(nsims = 10, x = RTL::fizdiffs, s0 = NULL)
}
\author{
Philippe Cote
}
