% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot2futConvergence}
\alias{spot2futConvergence}
\title{dataset: spot to futures convergence}
\format{
data frame
}
\source{
{Morningstar, EIA}
}
\usage{
spot2futConvergence
}
\description{
Cash and futures
}
\keyword{datasets}
