% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot2futCurve}
\alias{spot2futCurve}
\title{spot2futCurve}
\format{
data frame
}
\source{
{Morningstar, EIA}
}
\usage{
spot2futCurve
}
\description{
Forward Curve - cash and futures - demo dataset
}
\keyword{datasets}
