% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot2futCurve}
\alias{spot2futCurve}
\title{Data for spot to futures convergence - forward curve}
\format{
data frame
}
\source{
{Morningstar, EIA}
}
\usage{
spot2futCurve
}
\description{
Forward Curve
}
\keyword{datasets}
