% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot2futCurve}
\alias{spot2futCurve}
\title{dataset: spot to futures convergence curve}
\format{
data frame
}
\source{
{Morningstar, EIA}
}
\usage{
spot2futCurve
}
\value{
\code{tibble}
}
\description{
Forward Curve
}
\keyword{datasets}
