% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tickers_eia}
\alias{tickers_eia}
\title{Metadata of key EIA tickers grouped by products.}
\format{
data frame
}
\usage{
tickers_eia
}
\description{
Supports automated upload of EIA data through its API by categories.
Data frame organized by Supply Demand categories and products.
}
\keyword{datasets}
