% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeStrategyDY.R
\name{tradeStrategyDY}
\alias{tradeStrategyDY}
\title{Sample Quantitative Trading strategy}
\usage{
tradeStrategyDY(data = RTL::uso, par1value = 50, par2value = 200)
}
\arguments{
\item{data}{Dataframe of OHLC data e.g. RTL::uso}

\item{par1value}{Value of first parameter e.g. short MA}

\item{par2value}{Value of second parameter e.g. long MA}
}
\value{
Dataframe with indicators, signals, trades and profit and loss.
}
\description{
Based on dividend yield
}
\examples{
tradeStrategyDY(data = RTL::uso, par1value = 50, par2value = 200)
}
\author{
Philippe Cote
}
