% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usSwapCurves}
\alias{usSwapCurves}
\title{Data for US interest rate discounting using zero rates curve.}
\format{
List
#' @source Morningstar and FRED
}
\usage{
usSwapCurves
}
\description{
USD IR Discount, Forward and Zero curves from RQuantlib::DiscountCurve
}
\keyword{datasets}
