% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wtiSwap}
\alias{wtiSwap}
\title{dataset: WTI Calendar Month Average Swap pricing data}
\format{
data frame
}
\source{
{Morningstar}
}
\usage{
wtiSwap
}
\value{
\code{tibble}
}
\description{
WTI Crude futures
}
\keyword{datasets}
